% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfTools.R
\name{bin}
\alias{bin}
\title{Bin Manipulation in Maps}
\usage{
bin(obj, old, new = 1, extract = TRUE)
}
\arguments{
\item{obj}{\code{semiFrame} object.}

\item{old}{Integer, Character. Specifies bin or bins for manipulation. These bins are always cleared i.e. set to 0 after manipulation 
(except if \code{new} bin is in \code{old} bins).
Use keyword "all" for all the bins. Based on extract regime:
\itemize{
  \item{ \code{extract=TRUE} }{ all the bins will be cleared i.e. set to 0 after manipulation (except the \code{new} bin). }
  \item{ \code{extract=FALSE} }{ bins not specified in \code{old} will not be altered after manipulation. }
}}

\item{new}{Integer. Specifies single target bin. If the target bin is not present, it will be created (default: 1).}

\item{extract}{Logical. Relevant for \code{new > 0}. Provides automated housekeeping. If TRUE then all the bins (except target bin spcified in \code{new}) will be cleared i.e. set to 0 after the manipulation process (default: TRUE).}
}
\value{
Returns a \code{semiFrame} object with new bin values.
}
\description{
Bin manipulation in maps represented by a \code{semiFrame} object. In the case of Wafer Bin Maps
zero bin occurrences represent good dice. Any non zero bin represents a failed die.
}
\examples{

data(wbm)

# Create 1 fail map: Combine all the bins into bin 1
bin(wbm, "all", 1)

# Extract: Combine bins 2, 3, 5 into bin 1 and clear all other bins except bin 1
bin(wbm, c(2,3,5), 1)

# Merge: Combine bins 2, 3 into bin 2, clear bin 3 and retain all other bin occurrences
bin(wbm, c(2,3), 2, extract=FALSE)

# Erase: Clear bin 3 and retain all other bins
bin(wbm, 3, 0)

}
