% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfGraphics.R
\name{binPlot}
\alias{binPlot}
\title{Bin Pareto Plot}
\usage{
binPlot(obj, bins = "all", ..., main = NULL, clusterThr = 0.5, plot = TRUE)
}
\arguments{
\item{obj}{\code{semiFrame} object.}

\item{bins}{Integer. Specifies which bins to use. For all available bins use keyword "all" (default: "all").}

\item{...}{Cluster browsing. Input \code{cluster=clusterName} to browse directly into \code{clusterName} cluster.}

\item{main}{Character. Main title.}

\item{clusterThr}{Numeric. Threshold for cluster membership.}

\item{plot}{Logical. If \code{FALSE}, list of bin pareto percentages is returned for all maps and for each cluster (default: \code{TRUE}).}
}
\value{
Plot or list of tables of pareto percentages.
}
\description{
Plots contribution to yield loss by bins. If clusters are present, browsing is initiated.
}
\examples{

data(wbm)

# Plot bin pareto for all maps and all bins
binPlot(wbm)

# Plot bin pareto for all maps and bins 1, 2, 3
binPlot(wbm, bins=1:3)

# Browsing through the 'edge' cluster for all available bins
binPlot(wbm, bins="all", cluster="edge")

# Return yield loss contribution percentages table for all maps
binTab <- binPlot(wbm, plot=FALSE)$all

# Print all visual parameters
binPlotPar()

# Get bar border color
binPlotPar(colors.border)

# Set bar color to 'green'
binPlotPar(colors.box="green")

# Reload default visual parameters
binPlotPar(default)
}
