% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfGraphics.R
\name{binPlotPar}
\alias{binPlotPar}
\title{binPlot Parameters}
\usage{
binPlotPar(...)
}
\arguments{
\item{...}{Numeric, character or function. 
Root parameters:
\itemize{
  \item{ \code{colors} }{ Color profile. }
  \item{ \code{line} }{ Line attributes. }
  \item{ \code{space} }{ Space between bars (default: 0.2). }
}
\code{colors} parameters:
\itemize{
  \item{ \code{colors.box} }{ Bar color (default: "#80b1d3"). }
  \item{ \code{colors.border} }{ Bar border color. }
  \item{ \code{colors.line} }{ Cumulative line color (default: "#fb8072"). }
}
\code{line} parameters: 
\itemize{
  \item{ \code{line.lwd} }{ Cumulative line width (default: 2). }
}}
}
\description{
Gets, sets and resets visual aspects of binPlot.
}
\details{
Use parameter names to get values, use assignment to set parameter
values and use keyword \code{default} to load default values.
}
\examples{

data(wbm)

# Plot bin pareto for all maps and all bins
binPlot(wbm)

# Plot bin pareto for all maps and bins 1, 2, 3
binPlot(wbm, bins=1:3)

# Browsing through the 'edge' cluster for all available bins
binPlot(wbm, bins="all", cluster="edge")

# Return yield loss contribution percentages table for all maps
binTab <- binPlot(wbm, plot=FALSE)$all

# Print all visual parameters
binPlotPar()

# Get bar border color
binPlotPar(colors.border)

# Set bar color to 'green'
binPlotPar(colors.box="green")

# Reload default visual parameters
binPlotPar(default)
}
