% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfGraphics.R
\name{boxPlot}
\alias{boxPlot}
\title{Box Plot}
\usage{
boxPlot(
  obj,
  ...,
  outlier = NULL,
  xBr = 5,
  xBrMethod = "linear",
  xLevel = NULL,
  z = NULL,
  zBr = 10,
  zBrMethod = "yieldHeat",
  zLevel = NULL,
  zOutlier = NULL,
  main = NULL,
  numMain = TRUE,
  showLegend = TRUE,
  autoNorm = TRUE,
  useDataType = sfUseDataType,
  col = NULL
)
}
\arguments{
\item{obj}{\code{semiTable} or \code{semiFrame} object.}

\item{...}{Data columns to be used. Each column can be entered in following forms:
\itemize{
  \item{ integers } { Specify column indexes in the column group given by
    \code{useDataType} parameter. }
  \item{ characters }{ Specify column names. Both character vectors or strings without quotes
    can be used. } 
}}

\item{outlier}{Numeric. Outlier filtering of data column before plotting. Outliers are data points outside the
interval <(Q1-k1*IQR),(Q3+k2*IQR)>, where Q1 and Q3 are quartiles, IQR is interquartile range and
k1 and k2 are values specified by \code{outlier} parameter as its first and second element. 
If single value is provided, k1 = k2 setting is used. No outlier filtering is applied by default.}

\item{xBr}{Vector. Breaks of a numeric \code{x} column, works in conjunction with \code{xBrMethod} 
setting. See \code{zBr} for more information.}

\item{xBrMethod}{Character. Relevant for numeric \code{x} columns, works in conjunction with \code{xBr} setting. 
See \code{zBrMethod} for more information.}

\item{xLevel}{Vector. Non-null value indicates a categorical \code{x} column. Specifies which categories to use
for split. Use special value "all" to include all categories.}

\item{z}{Vector. Specifies column for \code{z} axis. This column can be broken into intervals 
using \code{zBr} and \code{zBrMethod} for numerical variables and using \code{zLevel} for 
categorical variables. Subsets of data column corresponding to \code{z} column breaks are 
then plotted with different colors. Input methods:
\itemize{
  \item{ integers } { Specify column indexes in the column group given by
    \code{useDataType} parameter. }
  \item{ characters }{ Specify column names. } 
} 
First found match is used when multivalued vector is provided.}

\item{zBr}{Vector. Breaks of a numeric \code{z} column, works in conjunction with \code{zBrMethod} 
setting. Single integer value determines number of breaks. Vector of \code{length>1} can be numerical,
character or POSIXct and represents interval breaking points. Special character values "min", "mean",
"median", "max" and "seq" can be used.}

\item{zBrMethod}{Character. Relevant for numeric \code{z} columns, works in conjunction with \code{zBr} setting. 
Split methods:
\itemize{
  \item{ "median" }{ Breaks \code{z} column with \code{median}. }
  \item{ "mean" }{ Breaks \code{z} column with \code{mean}. }
  \item{ "thr" }{ Breaks \code{z} column with threshold value specified by \code{zBr}. }
  \item{ "yieldHeat" }{ Corresponds to \code{zBr=c("seq(min, median, length.out=brValue)", "max")}, 
   where \code{brValue} is the value specified by \code{zBr}. }
  \item{ "linear" }{ Breaks \code{z} column into equidistant intervals, number of splits is given
   by \code{zBr}. }
  \item{ "hist" }{ Breaks \code{z} column using graphic's \code{hist} function. Values
   are passed to hist's \code{breaks} parameter using \code{zBr}. }
}}

\item{zLevel}{Vector. Non-null value indicates a categorical \code{z} column. Specifies which categories to use
for split. Use special value "all" to include all categories.}

\item{zOutlier}{Numeric. Outlier filtering of \code{z} column before plotting. For more details 
see \code{outlier} parameter.}

\item{main}{Character. Main title.}

\item{numMain}{Logical. When \code{TRUE} index of plotted column within corresponding \code{useDataType}
is included into the main title.}

\item{showLegend}{Logical. If \code{TRUE} legend is displayed.}

\item{autoNorm}{Logical. If \code{TRUE}, data prior normalization is restored for plotting.}

\item{useDataType}{Character. Specifies \code{obj} column group for integer column indexing.
First found match is used when multivalued vector is provided. 
See \code{\link{dataType}} for more information on data types.}

\item{col}{Character. Direct column input using character column names or integer column indices (default: \code{NULL}).}
}
\value{
Plot.
}
\description{
Plots box plot of a \code{semiTable} or \code{semiFrame} data column.
}
\examples{

data(pcy)

# Plot a numerical vs a categorical variable from the group
boxPlot(pcy, siteid, 1, xLevel="all")

# Plot a numerical vs a numerical variable from the group with default breaking
boxPlot(pcy, 2, 3)

\dontrun{# Various ways to input variables for plotting, first variable is used for x axis 
# Use console for browsing
var <- 10
boxPlot(pcy, 2, c(2,5), c(2:5), c(var), c("PARAM03"), 2:5, 8:-1, PARAM02, PARAM08:last)
}
# Break numerical x variable into arbitrary intervals
boxPlot(pcy, 2, 3, xBr=10)

# Use z variable for boxes coloring
boxPlot(pcy, 2, 3, z="ProbeYield", zBr=10)

# Filter outliers
boxPlot(pcy, wafer, UPYield, outlier=3)

# Print all visual parameters
boxPlotPar()
# Get chart margins
boxPlotPar(mar)
# Set relative width of the boxes
boxPlotPar(boxWex=0.2)
# Boxes width proportional to number of observations in the groups
boxPlotPar(varWidth=TRUE)
# Set legend position to top left
boxPlotPar(legend.pos="topleft")
# Reload default visual parameters
boxPlotPar(default)
}
\seealso{
\code{\link{dataType}} for available data types.
}
