% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semiFrame.R
\name{cbind.semiTable}
\alias{cbind.semiTable}
\alias{cbind}
\alias{cbind.semiFrame}
\title{Column-wise Binding}
\usage{
\method{cbind}{semiTable}(x, y)

\method{cbind}{semiFrame}(x, y)
}
\arguments{
\item{x}{\code{semiTable} or \code{semiFrame} object.}

\item{y}{\code{semiTable} or \code{semiFrame} object.}
}
\value{
A \code{semiTable} or \code{semiFrame} object with \code{tab} part columns binded.
}
\description{
Binds \code{tab} part columns of two \code{semiTable} and/or \code{semiFrame} objects.
}
\details{
Part \code{aid} and if applicable parts \code{mat} and \code{matDim} are all inherited from the \code{x} argument.
}
\examples{

data(pcy)

# Two semiTable objects with different PC parameters
pcyX <- rename(pcy, dataType(pcy)$PC, paste0("x_", dataType(pcy)$PC))
pcyY <- rename(pcy, dataType(pcy)$PC, paste0("y_", dataType(pcy)$PC))

# pcyX's and pcyY's PC parameters binded
pcyB <- cbind(pcyX, pcyY[,dataType(pcyY)$PC, with=FALSE])
 
}
