% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfModel.R
\name{cluster}
\alias{cluster}
\title{Cluster Management}
\usage{
cluster(obj, ..., add = FALSE, fromThr = 0.5, toVal = 1, verbose = TRUE)
}
\arguments{
\item{obj}{\code{semiTable} or \code{semiFrame} object.}

\item{...}{Parameters passed to provide bracket functionality for subsetting. In addition 3 parameters can be used:
\itemize{
  \item{ \code{from} }{ Specifies source cluster }
  \item{ \code{to} }{ Specifies target cluster }
  \item{ \code{remove} }{ Cluster for removal }
}}

\item{add}{Logical. Enables moving or copying of selected wafers, lots or wafers from a defined cluster 
into a cluster of choice. If \code{TRUE}, the selected group will be added (without alteration) to 
the specified cluster. If \code{FALSE}, the selected group will be moved to the specified cluster and 
removed from the original cluster (default: \code{FALSE}).}

\item{fromThr}{Numeric. Specifies the cluster threshold and expected in the [0,1] 
range. Only cluster group members with presence value bigger than \code{fromThr} will 
be included (default: 0.5).}

\item{toVal}{Numeric. Specifies target cluster value describing cluster presence. 
Expected in the [0,1] range. (default: 1)}

\item{verbose}{Logical. If \code{FALSE} turns off function messages (default: \code{TRUE}).}
}
\value{
\code{semiTable} or \code{semiFrame} object
}
\description{
Cluster management including: creation of new clusters, removing existing clusters, 
moving/coping of specific wafers, lots or wafers belonging to a certain cluster to another cluster.
}
\examples{

data(wbm)

# Move all wafers from cluster clear to cluster center
cluster(wbm, from=clear, to=center)

# Add all wafers from cluster edge to cluster center
cluster(wbm, from=edge, to=center, add=TRUE)

# Add lot LOT003 to cluster clear
cluster(wbm, lot=="LOT003", to=clear, add=TRUE)

# Add wafer #9 from lot LOT002 to cluster center
cluster(wbm, lotWafer=="LOT002-W09", to=center, add=TRUE)

# Create a new cluster, include lot LOT005 with probeEndDate before 5th March 2016
cluster(wbm, lot=="LOT005"&probeEndDate<"2016-03-05", to=newCluster)

# Remove cluster edge
cluster(wbm, remove=edge)

}
