% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfTools.R
\name{colCor}
\alias{colCor}
\title{Columns Correlation}
\usage{
colCor(
  obj,
  col1,
  col2,
  by = "lotWafer",
  outputOrder = "cint",
  alternative = "g",
  confLevel = 0.95
)
}
\arguments{
\item{obj}{\code{semiTable} or \code{semiFrame} object.}

\item{col1}{Character. Specifies the first column name (of a pair) for mutual 
correlation assessment.}

\item{col2}{Character. Specifies the second column name (of a pair) for mutual 
correlation assessment.}

\item{by}{Character. Defines grouping in the output table (default: "lotWafer").}

\item{outputOrder}{Character. Defines sorting column in the output table:
\itemize{ 
  \item{ "gcor" }{ Grouped correlation value. }
  \item{ "pval" }{ p-value for given specified \code{alternative}. } 
  \item{ "cint" }{ Left bound of confidence interval given \code{confLevel} (default). }
}}

\item{alternative}{Character. Indicates the alternative hypothesis and must be one of
"two.sided", "greater" or "less" (\code{pval} in the output table) (default: "g").}

\item{confLevel}{Numeric. Confidence level for the returned confidence interval (\code{cint} in the output table) (default: 0.95).}
}
\value{
A \code{data.table} with correlation statistics.
}
\description{
Returns correlation between two columns with grouping.
}
\examples{

data(pcy)

# Introduce a new variable myPARAM
pcy2 <- variable(pcy, "myPARAM", 2*pcy[,PARAM02]-pcy[,PARAM03]/2, "PC")
# Find out correlation of the new variable with PARAM10
colCor(pcy2, "myPARAM", "PARAM10", by = "lotWafer")

}
