% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfModel.R
\name{completeImpute}
\alias{completeImpute}
\title{Completes semiTable from semiImpute}
\usage{
completeImpute(siObj, omitNA = FALSE)
}
\description{
Completes semiTable from semiImpute, returns list of semiTables
}
