% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfGraphics.R
\name{corPlot}
\alias{corPlot}
\title{Correlations Plot}
\usage{
corPlot(
  obj,
  ...,
  chart = "mat",
  addYield = "ProbeYield",
  matLimit = 0,
  matMethod = "color",
  matType = "lower",
  matDiag = FALSE,
  matOrder = "hclust",
  hclustLinkage = "complete",
  dendThr = 0.9,
  dendMeasure = "abs",
  main = NULL,
  useDataType = sfUseDataType,
  col = NULL
)
}
\arguments{
\item{obj}{\code{semiTable} or \code{semiFrame} object.}

\item{...}{Data columns to be used. If none are specified all columns from the data type
given by \code{useDataType} will be used. Each column can be entered in following forms:
\itemize{
  \item{ integers } { Specify column indexes in the column group given by
    \code{useDataType} parameter. }
  \item{ characters }{ Specify column names. Both character vectors or strings without quotes
    can be used. } 
}}

\item{chart}{Character. Visualization type:
\itemize{
  \item{ "mat" }{ for correlation matrix (default). }
  \item{ "dend" }{ for correlation dendrogram. }
}}

\item{addYield}{Character. Specifies yield column to add (default: "ProbeYield").}

\item{matLimit}{Numeric. Relevant for \code{chart="mat"}. Only correlations with absolute values above this threshold are shown (default: 0).}

\item{matMethod}{Character. Relevant for \code{chart="mat"}. Supported visualization methods: "circle", "square", "ellipse", "number",
"pie", "shade" and "color" (default).}

\item{matType}{Character. Relevant for \code{chart="mat"}. Types are "full", "upper" or "lower" (default), for full matrix, 
upper triangular or lower triangular matrix, respectively.}

\item{matDiag}{Logical. Relevant for \code{chart="mat"}. TRUE to add diagonal (default: FALSE).}

\item{matOrder}{Character. Relevant for \code{chart="mat"}. Ordering method of the correlation matrix.
\itemize{
  \item{ "hclust" }{ hierarchical clustering order (default). }
  \item{ "original" }{ original order. }
  \item{ "AOE" }{ the angular order of the eigenvectors. }
  \item{ "FPC" }{ first principal component order. }
  \item{ "alphabet" }{ alphabetical order. }
}}

\item{hclustLinkage}{Character. Relevant for \code{chart="dend"}. Agglomeration method for dendrogram:
"ward.D", "ward.D2", "single", "complete" (default), "average", "mcquitty", "median" or "centroid".}

\item{dendThr}{Numeric. Relevant for \code{chart="dend"}. Threshold for red boxes (default: 0.9).}

\item{dendMeasure}{Character. Relevant for \code{chart="dend"}. Choose one of "pos", "neg" and "abs". Then zero
distance in dendogram represents correlation +1,-1 and -1 or +1, respectively (default: "abs").}

\item{main}{Character. Main title.}

\item{useDataType}{Character. Specifies \code{obj} column group for integer column indexing.
First found match is used when multivalued vector is provided. 
See \code{\link{dataType}} for more information on data types.}

\item{col}{Character. Direct column input using character column names or integer column indices (default: \code{NULL}).}
}
\value{
Plot.
}
\description{
Plots correlation matrix or dendogram.
}
\examples{

data(pcy)

# Plotting correlation matrix
corPlot(pcy)

# Plot dendogram with custom grouping threshold
corPlot(pcy, chart="dend", dendThr=0.95)

# Plot upper triangle correlation matrix and show correlations above 0.5 only
corPlot(pcy, chart="mat", matLimit=0.5, matType="upper")

# Use single linkage for dendrogram
corPlot(pcy, chart="dend", hclustLinkage="single")

# Print all visual parameters
corPlotPar()
# Get grouping rectangle color
corPlotPar(dend.thrColor)
# Set grouping rectangle color
corPlotPar(dend.thrColor="green")
corPlot(pcy, chart="dend", main="Threshold color changed")

# Reload default visual parameters
corPlotPar(default)
}
