% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semiFrame.R
\name{dataType.semiFrame}
\alias{dataType.semiFrame}
\alias{dataType}
\alias{dataType<-.semiFrame}
\alias{dataType.semiTable}
\alias{dataType<-.semiTable}
\title{Manage Data Types}
\usage{
\method{dataType}{semiFrame}(obj)

\method{dataType}{semiFrame}(obj) <- value

\method{dataType}{semiTable}(obj)

\method{dataType}{semiTable}(obj) <- value
}
\arguments{
\item{obj}{\code{semiTable} or \code{semiFrame} object.}
}
\value{
A list of data types with corresponding variables.
}
\description{
Gets or sets \code{dataType} attribute of a \code{semiTable} or a \code{semiFrame} object.
}
\details{
Data types stored in the \code{dataType} attribute represent groups of variables.
If applicable, the following data types can be used:
\itemize{
  \item{ "META" }{ for Meta parameters. }
  \item{ "PC" }{ for PCM parameters. }
  \item{ "PCJ" }{ for PCM Junk parameters. }
  \item{ "MET" }{ for Metrology parameters. }
  \item{ "LEH" }{ for Lot Equipment History parameters. }
  \item{ "DEF" }{ for Defectivity parameters. }
  \item{ "Y" }{ for Yield variables. }
  \item{ "C" }{ for Clusters variables. }
  \item{ "FG" }{ for Fab Guard parameters. }
  \item{ "DR" }{ for Dimensionality Reduction variables. }
  \item{ "U" }{ for User-defined variables. }
}
}
\examples{
 
data(pcy)

# Get the dataType attribute of the pcy object
dataType(pcy)

# Show data types
names(dataType(pcy))

# Show variable names from the 'PC' data type
dataType(pcy)$PC

}
