% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfGraphics.R
\name{defBayesOptimPlotPar}
\alias{defBayesOptimPlotPar}
\title{Bayes Defectivity Optimization Plot Parameters}
\usage{
defBayesOptimPlotPar(...)
}
\arguments{
\item{...}{Numeric, character or function. 

Root parameters:
\itemize{
  \item{ \code{color} }{ Color parameters. }
}
\code{color} parameters:
\itemize{
  \item{ \code{color.yield} }{ Color setting for the "yield" chart type (default: "#80B1D3").}
  \item{ \code{color.prob} }{ Color setting for the "prob" chart type.}
}}
}
\description{
Gets, sets and reloads default visual parameters of defBayesOptimPlot.
}
\details{
Use parameter names to get values, use assignment to set parameter
values and use keyword \code{default} to load default values.
}
\examples{

##############################################
# Examples for defBayesOptimPlot - To be added
##############################################

# Print all visual parameters
defBayesOptimPlotPar()
     
# Set color setting for the "yield" chart to red
defBayesOptimPlotPar(color.yield="red")

# Get color setting for the "prob" chart
defBayesOptimPlotPar(color.prob)     

# Reload default visual parameters
defBayesOptimPlotPar(default)
}
