% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semiFrame.R
\name{defDie}
\alias{defDie}
\title{Defectivity per Die}
\usage{
defDie(
  wbmObj,
  demObj,
  wbmAlign = NULL,
  demAlign = NULL,
  NAtoZero = FALSE,
  def = NULL,
  ded = NULL,
  core = 8,
  verbose = TRUE
)
}
\arguments{
\item{wbmObj}{`semiFrame` object representing the wafer bin map dataset.}

\item{demObj}{`semiFrame` object representing the defect maps dataset.}

\item{wbmAlign}{Integer. User defined alignment of wafer bin maps. 
First to the fourth element corresponds to the top, right, bottom, left side of the map (default: \code{NULL}).}

\item{demAlign}{Integer. User defined alignment of defect maps. 
First to the fourth element corresponds to the top, right, bottom, left side of the map (default: \code{NULL}).}

\item{NAtoZero}{Logical. If \code{TRUE}, all the NAs will be set to zero (default: \code{FALSE}).}

\item{def}{`semiFrame` object representing defects dataset. Will be used to fill inheriting information if provided. (default: \code{NULL}).}

\item{ded}{`semiFrame` object representing defects per die dataset from previous run of \code{defDie} (default: \code{NULL}).}

\item{core}{Integer. Specifies number of cores for parallel computation.}

\item{verbose}{Logical. If \code{FALSE} turns off function messages (default: \code{TRUE}).}
}
\value{
Die-level \code{semiTable} object. Column \code{ALL} represents defect count on the die.
}
\description{
Matches defect maps with wafer bin maps and creates die-level table ie. each row represents a die.
}
\examples{

# Defects dataset
data(def)

# Defect maps dataset
data(dem)

# Defects per die dataset
ded <- defDie(wbm, dem, def=def)

}
