% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semiFrame.R
\name{defInherited}
\alias{defInherited}
\title{Defects Inheriting}
\usage{
defInherited(obj, steps, type = "auto", verbose = TRUE)
}
\arguments{
\item{obj}{\code{semiTable} object. Applicable to defects dataset or to defects per die dataset.}

\item{steps}{Character. Defines manufacturing steps with correct order.}

\item{type}{Character. Specifies the type of dataset: "auto" (default),
"def" corresponding to defects dataset or "die" corresponding to defects per die dataset.}

\item{verbose}{Logical. If \code{FALSE} turns off function messages (default: \code{TRUE}).}
}
\value{
Percentages of defects inherited or percentages of dice that inherit one or more defect.
}
\description{
Computes percentages of defects inherited from previous manufacturing steps from defect inspection data.
}
\examples{

data(wbm)

# Defects dataset
data(def)
# Defect maps dataset
data(dem)
# Defects per die dataset
ded <- defDie(wbm, dem, def=def)

# Percentages of defects inherited
defInherited(def, steps=c("STEP1", "STEP2"))
# Percentages of dice that inherit one or more defect
defInherited(ded, steps=c("STEP1", "STEP2"))

}
