% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfGraphics.R
\name{defInheritedPlot}
\alias{defInheritedPlot}
\title{Inherited Defects Percentages Plot}
\usage{
defInheritedPlot(
  obj,
  steps,
  col = "DefArea",
  breaks = c(0:10, +Inf),
  core = (length(breaks) - 1),
  main = NULL,
  showLegend = TRUE
)
}
\arguments{
\item{obj}{\code{semiTable} object. Applicable to defects dataset.}

\item{steps}{Character. Defines manufacturing steps with correct order.}

\item{col}{Character. Specifies column to break into intervals (default: "DefArea").}

\item{breaks}{Numeric. Cut points of intervals of the variable specified in \code{col} (default: \code{c(0:10, +Inf)}).}

\item{core}{Integer. Specifies number of cores for parallel computation.}

\item{main}{Character. Main title.}

\item{showLegend}{Logical. If \code{TRUE} legend is displayed.}
}
\value{
Plot.
}
\description{
Plots inherited defects percentages per specified intervals.
}
\examples{

data(def)

# Create inherited defects percentages plot
defInheritedPlot(def, steps=c("STEP1", "STEP2"))

# Print all visual parameters
defInheritedPlotPar()

# Set legend position to topleft
defInheritedPlotPar(legend.pos="topleft")

# Reload default visual parameters
defInheritedPlotPar(default)
}
