% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfGraphics.R
\name{defInheritedPlotPar}
\alias{defInheritedPlotPar}
\title{Inherited Defects Percentages Plot Parameters}
\usage{
defInheritedPlotPar(...)
}
\arguments{
\item{...}{Numeric, character or function. 

Root parameters:
\itemize{
  \item{ \code{level} }{ Color scheme. }
  \item{ \code{lwd} }{ Line width (default: 2.5). }
  \item{ \code{legend} }{ Legend parameters. }
}
\code{legend} parameters:
\itemize{
  \item{ \code{legend.pos} }{ Legend position (default: "bottomright").}
}}
}
\description{
Gets, sets and reloads default visual parameters of defInheritedPlot.
}
\details{
Use parameter names to get values, use assignment to set parameter
values and use keyword \code{default} to load default values.
}
\examples{

data(def)

# Create inherited defects percentages plot
defInheritedPlot(def, steps=c("STEP1", "STEP2"))

# Print all visual parameters
defInheritedPlotPar()

# Set legend position to topleft
defInheritedPlotPar(legend.pos="topleft")

# Reload default visual parameters
defInheritedPlotPar(default)
}
