% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfTools.R
\name{degrade}
\alias{degrade}
\title{Degrade to Lower Level}
\usage{
degrade(obj, degradeTo, fun = median)
}
\arguments{
\item{obj}{\code{semiTable} or \code{semiFrame} object.}

\item{degradeTo}{Character. Specifies the level to which the data should be degraded to: 
\itemize{
  \item{ "wafer" }{ for wafer level degradation. }
  \item{ "lot" }{ for lot level degradation. }
}}

\item{fun}{Function. Specifies aggregation function. For example \code{median} (default) or \code{mean}.}
}
\value{
\code{semiTable} or \code{semiFrame} object
}
\description{
Degrades to lot or wafer level and aggregates column values.
}
\examples{

data(pcy)

# Degrade site PCM data to wafer level
degrade(pcy, "wafer")

# Degrade site PCM or wafer level data to lot level
degrade(pcy, "lot")

}
