% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfTools.R
\name{deNorm}
\alias{deNorm}
\title{Restore Values}
\usage{
deNorm(obj, verbose = TRUE)
}
\arguments{
\item{obj}{\code{semiTable} or \code{semiFrame} object.}

\item{verbose}{Logical. If \code{FALSE} turns off function messages (default: \code{TRUE}).}
}
\value{
Copy of the \code{semiTable} or \code{semiFrame} object with restored values.
}
\description{
Restore values prior normalization if relevant.
}
\examples{

data(pcy)

# Restore values
deNorm(pcy)

}
