% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfGraphics.R
\name{distrShiftPlot}
\alias{distrShiftPlot}
\title{Distribution Shifting and Narrowing Plot}
\usage{
distrShiftPlot(
  obj,
  ...,
  chart = "density",
  xCenter = NULL,
  xNarrow = NULL,
  xCenterFun = median,
  xSizeRatio = ifelse(is.null(xNarrow), 0.5, xNarrow),
  z = NULL,
  zBr = 2,
  zBrMethod = "median",
  zLevel = NULL,
  zCenter = NULL,
  zNarrow = NULL,
  zCenterFun = median,
  zSizeRatio = ifelse(is.null(zNarrow), 0.5, zNarrow),
  freq = FALSE,
  main = NULL,
  showLegend = TRUE,
  zShowLegend = TRUE,
  autoNorm = TRUE,
  useDataType = sfUseDataType,
  plot = TRUE,
  col = NULL
)
}
\arguments{
\item{obj}{\code{semiTable} or \code{semiFrame} object.}

\item{...}{Data columns to be used as \code{x} variable. Each column can be entered in following forms:
\itemize{
  \item{ integers } { Specify column indexes in the column group given by
    \code{useDataType} parameter. }
  \item{ characters }{ Specify column names. Both character vectors or strings without quotes
    can be used. } 
}}

\item{chart}{Character. Type of chart:
\itemize{
  \item{ "hist" } { Plots histogram layers. }
  \item{ "density" } { Plots density layers (default). }
}}

\item{xCenter}{Numeric. Centers of \code{x} variable's shifted distribution. If more than one center point
is provided browsing mode will be initiated (default: \code{NULL}).}

\item{xNarrow}{Numeric. Narrowing coefficient for \code{x} variable's distribution.}

\item{xCenterFun}{Character. Center calculation function for \code{x} variable's  distribution. Examples: \code{median}, \code{mean}.}

\item{xSizeRatio}{Numeric. Proportion of the data used for sampling \code{x} variable's distribution.}

\item{z}{Vector. Specifies column for \code{z} axis. This column can be broken into intervals 
using \code{zBr} and \code{zBrMethod} for numerical variables and using \code{zLevel} for 
categorical variables. Subsets of data column corresponding to \code{z} column breaks are 
then plotted with different colors. Input methods:
\itemize{
  \item{ integers } { Specify column indexes in the column group given by
    \code{useDataType} parameter. }
  \item{ characters }{ Specify column names. } 
} 
First found match is used when multivalued vector is provided.}

\item{zBr}{Vector. Breaks of a numeric \code{z} column, works in conjunction with \code{zBrMethod} 
setting. Single integer value determines number of breaks. Vector of \code{length>1} can be numerical,
character or POSIXct and represents interval breaking points. Special character values "min", "mean",
"median", "max" and "seq" can be used.}

\item{zBrMethod}{Character. Relevant for numeric \code{z} columns, works in conjunction with \code{zBr} setting. 
Split methods:
\itemize{
  \item{ "median" }{ Breaks \code{z} column with \code{median}. }
  \item{ "mean" }{ Breaks \code{z} column with \code{mean}. }
  \item{ "thr" }{ Breaks \code{z} column with threshold value specified by \code{zBr}. }
  \item{ "yieldHeat" }{ Corresponds to \code{zBr=c("seq(min, median, length.out=brValue)", "max")}, 
   where \code{brValue} is the value specified by \code{zBr}. }
  \item{ "linear" }{ Breaks \code{z} column into equidistant intervals, number of splits is given
   by \code{zBr}. }
  \item{ "hist" }{ Breaks \code{z} column using graphic's \code{hist} function. Values
   are passed to hist's \code{breaks} parameter using \code{zBr}. }
}}

\item{zLevel}{Vector. Non-null value indicates a categorical \code{z} column. Specifies which categories to use
for split. Use special value "all" to include all categories.}

\item{zCenter}{Numeric. Centers of \code{z} variable's shifted distribution.}

\item{zNarrow}{Numeric. Narrowing coefficient for \code{z} variable's distribution.}

\item{zCenterFun}{Character. Center calculation function for \code{z} variable's  distribution. Examples: \code{median}, \code{mean}.}

\item{zSizeRatio}{Numeric. Proportion of the data used for sampling \code{z} variable's distribution.}

\item{freq}{Logical. When \code{TRUE}, \code{y} axis provides density, frequency otherwise (default: \code{FALSE}).}

\item{main}{Character. Main title.}

\item{showLegend}{Logical. If \code{TRUE} legend is displayed.}

\item{zShowLegend}{Logical. When \code{TRUE} legend corresponding to \code{z} variable is displayed (default: \code{TRUE}).}

\item{autoNorm}{Logical. If \code{TRUE}, data prior normalization is restored for plotting.}

\item{useDataType}{Character. Specifies \code{obj} column group for integer column indexing.
First found match is used when multivalued vector is provided. 
See \code{\link{dataType}} for more information on data types.}

\item{plot}{Logical. When \code{FALSE}, function returns shifted distribution's data (default: \code{TRUE}).}

\item{col}{Character. Direct column input using character column names or integer column indices (default: \code{NULL}).}
}
\value{
Plot by default, if \code{plot=FALSE} function returns shifted distribution's data.
}
\description{
Plots shifted and narrowed distribution of a variable possibly depending on \code{z} variable's
breaks or shifting and narrowing as well.
}
\examples{

data(pcy)

pcy <- model(pcy, type="class", model="forest", splitRatio=1, formula="ProbeYield~.")

# Distribution shifting and narrowing visualization
distrShiftPlot(pcy, PARAM01, xCenter=22, xNarrow=0.5)
\dontrun{
# Distribution shifting and narrowing, browse through centers
distrShiftPlot(pcy, PARAM01, xCenter=c(22,23), xNarrow=0.5)
}

\dontrun{
# Select numerical z variable with input by index
distrShiftPlot(pcy, PARAM01, xCenter=22, z=5)
# Select categorical z variable with input by name
distrShiftPlot(pcy, 1, xCenter=22, z="part", zLevel=c("PART1", "PART2"))
}
# Special keyword "all" for all levels
distrShiftPlot(pcy, 1, xCenter=22, z="part", zLevel="all")

# Print all visual parameters
distrShiftPlotPar()
# Get distribution's colors
distrShiftPlotPar(colors.distr)
# Set legend position to bottomleft
distrShiftPlotPar(legend.pos="bottomleft")

# Reload default visual parameters
distrShiftPlotPar(default)
}
