% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfGraphics.R
\name{distrShiftPlotPar}
\alias{distrShiftPlotPar}
\title{Distribution Shift Plot Parameters}
\usage{
distrShiftPlotPar(...)
}
\arguments{
\item{...}{Numeric, character or function. 

Root parameters
\itemize{
 \item{ \code{colors} }{ Color schemes. }
 \item{ \code{signif} }{ Number of significant digits (default: 3). }
 \item{ \code{pbr} }{ Default number of histogram cells for "hist" splits (default: 60). }
 \item{ \code{space} }{ Space between bars (default: 0). }
 \item{ \code{legend} }{ Legend parameters. }
 \item{ \code{zLegend} }{ Legend corresponding do \code{z} variable. }
}
\code{colors} parameters:
\itemize{
  \item{ \code{colors.alpha} }{ Transparency setting (default: 0.7). }
  \item{ \code{colors.distr} }{ Distributions colors (default: "#BDBDBDB3", "#8DD3C7B3"). }
  \item{ \code{colors.border} }{ Distributions borders colors (default: "#AAAAAAFF", "#7FBEB3FF"). }
}
\code{legend} parameters:
\itemize{
  \item{ legend.pos }{ Legend position (default: "topright"). }
  \item{ \code{legend.lty} }{ Lines type of the legend (default: 0). }
  \item{ \code{legend.lwd} }{ Lines width of the legend (default: 1). }
  \item{ \code{legend.cex} }{ Annotation magnification of the legend (default: 0.9). }
  \item{ \code{legend.bty} }{ Type of box that is drawn around the legend (default: "n"). }
  \item{ \code{legend.segLen} }{ The length of lines drawn to illustrate ‘lty’ and/or ‘lwd’
          (in units of character widths) (default: "0"). }
  \item{ \code{legend.pch} }{ Points symbols of the legend (default: 15). }
  \item{ \code{legend.ptCex} }{ A numeric determining the size of the exemplary vertex in the legend box (default: 1.8). }
}
\code{zLegend} parameters:
\itemize{
  \item{ zLegend.pos }{ Legend position (default: "topleft"). }
}}
}
\description{
Gets, sets and reloads default visual parameters of distrShiftPlot.
}
\details{
Use parameter names to get values, use assignment to set parameter
values and use keyword \code{default} to load default values.
}
\examples{

data(pcy)

pcy <- model(pcy, type="class", model="forest", splitRatio=1, formula="ProbeYield~.")

# Distribution shifting and narrowing visualization
distrShiftPlot(pcy, PARAM01, xCenter=22, xNarrow=0.5)
\dontrun{
# Distribution shifting and narrowing, browse through centers
distrShiftPlot(pcy, PARAM01, xCenter=c(22,23), xNarrow=0.5)
}

\dontrun{
# Select numerical z variable with input by index
distrShiftPlot(pcy, PARAM01, xCenter=22, z=5)
# Select categorical z variable with input by name
distrShiftPlot(pcy, 1, xCenter=22, z="part", zLevel=c("PART1", "PART2"))
}
# Special keyword "all" for all levels
distrShiftPlot(pcy, 1, xCenter=22, z="part", zLevel="all")

# Print all visual parameters
distrShiftPlotPar()
# Get distribution's colors
distrShiftPlotPar(colors.distr)
# Set legend position to bottomleft
distrShiftPlotPar(legend.pos="bottomleft")

# Reload default visual parameters
distrShiftPlotPar(default)
}
