% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semiFrame.R
\name{getRowRas.semiFrame}
\alias{getRowRas.semiFrame}
\title{Get the matrix corresponding to a row}
\usage{
\method{getRowRas}{semiFrame}(obj, i = NULL, repo = NULL)
}
\description{
semiFrame method to get the matrix corresponding to a row
}
