% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfModel.R
\name{heat}
\alias{heat}
\title{Create Cluster Heatmap}
\usage{
heat(obj, thr = 0.5)
}
\arguments{
\item{obj}{\code{semiTable} or \code{semiFrame} object.}

\item{thr}{Numeric. Specifies the cluster threshold and should be in the (0,1) range. Only cluster group members with
presence value bigger than \code{thr} will be included in the heatmap.}
}
\value{
\code{semiFrame} object
}
\description{
Creates heatmaps by averaging provided maps for each cluster.
}
\examples{

data(wbm)

# Create heatmaps of clusters for all bins
heat(bin(wbm, "all"))

}
