% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfGraphics.R
\name{histPlot}
\alias{histPlot}
\title{Histogram Plot}
\usage{
histPlot(
  obj,
  ...,
  xBr = NULL,
  xBrMethod = "hist",
  xLevel = NULL,
  outlier = NULL,
  z = c("ProbeYield", "UPYield"),
  zBr = 2,
  zBrMethod = "yieldHeat",
  zLevel = NULL,
  zOutlier = NULL,
  chart = "density",
  yLim = NULL,
  main = NULL,
  numMain = TRUE,
  showLegend = TRUE,
  autoNorm = TRUE,
  useDataType = sfUseDataType,
  col = NULL
)
}
\arguments{
\item{obj}{\code{semiTable} or \code{semiFrame} object.}

\item{...}{Variables (data columns) to be plotted. Each column can be entered in following forms:
\itemize{
  \item{ integers } { Specify column indexes in the column group given by
    \code{useDataType} parameter. }
  \item{ characters }{ Specify column names. Both character vectors or strings without quotes
    can be used. } 
} 
Browsing mode is activated if multiple columns are provided.}

\item{xBr}{Vector. Breaks of a numeric \code{x} column, works in conjunction with \code{xBrMethod} 
setting. See \code{zBr} for more information.}

\item{xBrMethod}{Character. Relevant for numeric \code{x} columns, works in conjunction with \code{xBr} setting. 
See \code{zBrMethod} for more information.}

\item{xLevel}{Vector. Non-null value indicates a categorical \code{x} column. Specifies which categories to use
for split. Use special value "all" to include all categories.}

\item{outlier}{Numeric. Outlier filtering of data column before plotting. Outliers are data points outside the
interval <(Q1-k1*IQR),(Q3+k2*IQR)>, where Q1 and Q3 are quartiles, IQR is interquartile range and
k1 and k2 are values specified by \code{outlier} parameter as its first and second element. 
If single value is provided, k1 = k2 setting is used. No outlier filtering is applied by default.}

\item{z}{Vector. Specifies column for \code{z} axis. This column can be broken into intervals 
using \code{zBr} and \code{zBrMethod} for numerical variables and using \code{zLevel} for 
categorical variables. Subsets of data column corresponding to \code{z} column breaks are 
then plotted with different colors. Input methods:
\itemize{
  \item{ integers } { Specify column indexes in the column group given by
    \code{useDataType} parameter. }
  \item{ characters }{ Specify column names. } 
} 
First found match is used when multivalued vector is provided.}

\item{zBr}{Vector. Breaks of a numeric \code{z} column, works in conjunction with \code{zBrMethod} 
setting. Single integer value determines number of breaks. Vector of \code{length>1} can be numerical,
character or POSIXct and represents interval breaking points. Special character values "min", "mean",
"median", "max" and "seq" can be used.}

\item{zBrMethod}{Character. Relevant for numeric \code{z} columns, works in conjunction with \code{zBr} setting. 
Split methods:
\itemize{
  \item{ "median" }{ Breaks \code{z} column with \code{median}. }
  \item{ "mean" }{ Breaks \code{z} column with \code{mean}. }
  \item{ "thr" }{ Breaks \code{z} column with threshold value specified by \code{zBr}. }
  \item{ "yieldHeat" }{ Corresponds to \code{zBr=c("seq(min, median, length.out=brValue)", "max")}, 
   where \code{brValue} is the value specified by \code{zBr}. }
  \item{ "linear" }{ Breaks \code{z} column into equidistant intervals, number of splits is given
   by \code{zBr}. }
  \item{ "hist" }{ Breaks \code{z} column using graphic's \code{hist} function. Values
   are passed to hist's \code{breaks} parameter using \code{zBr}. }
}}

\item{zLevel}{Vector. Non-null value indicates a categorical \code{z} column. Specifies which categories to use
for split. Use special value "all" to include all categories.}

\item{zOutlier}{Numeric. Outlier filtering of \code{z} column before plotting. For more details 
see \code{outlier} parameter.}

\item{chart}{Character. Type of chart:
\itemize{
  \item{ "hist" } { Plots histogram layers. }
  \item{ "density" } { Plots density layers (default). }
  \item{ "beside" }{ Plots histogram groups for each bin. } 
}}

\item{yLim}{Numeric. User defined y axis limits. Vector of length 2.}

\item{main}{Character. Main title.}

\item{numMain}{Logical. When \code{TRUE} index of plotted column within corresponding \code{useDataType}
is included into the main title.}

\item{showLegend}{Logical. If \code{TRUE} legend is displayed.}

\item{autoNorm}{Logical. If \code{TRUE}, data prior normalization is restored for plotting.}

\item{useDataType}{Character. Specifies \code{obj} column group for integer column indexing.
First found match is used when multivalued vector is provided. 
See \code{\link{dataType}} for more information on data types.}

\item{col}{Character. Direct column input using character column names or integer column indices (default: \code{NULL}).}
}
\value{
Plot.
}
\description{
Plots histogram of a \code{semiTable} or \code{semiFrame} variable (data column). 
Chart colors are based on \code{z} axis column.
}
\examples{

data(pcy)

# Plot first variable from the group
histPlot(pcy, 1)

\dontrun{# Various ways to input variables for plotting 
# Use console for browsing
var <- 10
histPlot(pcy, 1, c(1,5), c(1:5), c(var), c("PARAM03"), 1:5, 8:-1, PARAM01, PARAM08:last)
}
# Select numerical z variable with input by index
histPlot(pcy, UPYield, z=5)
# Select categorical z variable with input by name
histPlot(pcy, 1, z="part", zLevel=c("PART1", "PART2"))
# Special keyword "all" for all levels
histPlot(pcy, 1, z="part", zLevel="all")
# Select date type z variable
histPlot(pcy, 1, z="pcStartDate", zBr=as.POSIXct(c("2015/10/01", "2016/01/01", "2016/10/01")))

# Filter outliers
histPlot(pcy, UPYield, outlier=10, z=5, zOutlier=3)

# Break numerical z variables into intervals
# No break - single color mode
histPlot(pcy, 1, z=NULL)
# No break
histPlot(pcy, 1, zBr=1)
# Breaking by median, mean and yieldHeat method
histPlot(pcy, 1, zBrMethod="median")
histPlot(pcy, 1, zBrMethod="mean")
histPlot(pcy, 1, zBr=5, zBrMethod="yieldHeat")
# br keywords seq, min, median, mean and max together with values
histPlot(pcy, 10, chart="hist", zBr=c("min", "seq(20, 80, length.out=3)", "max"))
# Equivalent for zBrMethod="yieldHeat" using zBr keywords
histPlot(pcy, 1, zBr=c("seq(min, median, length.out=5)", "max"))

# Categorical x variable
histPlot(pcy, wafer, xLevel="all", z=NULL)

# Hide legend
histPlot(pcy, 1, showLegend=FALSE)

# Print all visual parameters
histPlotPar()
# Get value of a visual parameter
histPlotPar(colors.zBr)
# Set breaking colors to red and green, transparency to zero and legend position to 
# top left
histPlotPar(colors.zBr=c("red", "green"), colors.alpha.density=1, legend.pos="topleft")

# Reload default visual parameters
histPlotPar(default)
}
