% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfLoad.R
\name{loadData}
\alias{loadData}
\alias{saveData}
\title{Load and Save Data}
\usage{
loadData(
  fileName,
  tech = NULL,
  repo = "dev",
  ver = "latest",
  subVer = NULL,
  user = NULL,
  path = NULL,
  ext = "rda",
  verbose = TRUE
)

saveData(
  obj,
  tech = NULL,
  repo = "dev",
  ver = "latest",
  subVer = NULL,
  user = NULL,
  path = NULL,
  fileName = NULL,
  ext = "rda",
  envir = .GlobalEnv,
  verbose = TRUE
)
}
\arguments{
\item{fileName}{Character. Specifies name of the file to be loaded or saved.}

\item{tech}{Character. Specifies name of the technology corresponding to the data (default: \code{NULL}).}

\item{repo}{Character. Defines the name of the data repository (default: "dev").}

\item{ver}{Character. Data version. Use keyword "latest" for latest version available. (default: "latest").}

\item{subVer}{Character. Additional information on data version (default: \code{NULL}).}

\item{user}{Character. Specifies the \code{inLab} account username (default: \code{NULL}).}

\item{path}{Character. Use to enter complete file path instead of \code{tech}, \code{repo}, \code{ver}, \code{subVer} combination (default: \code{NULL}).}

\item{ext}{Character. Sets the file extension (default: "rda").}

\item{verbose}{Logical. If \code{FALSE} turns off function messages (default: \code{TRUE}).}

\item{obj}{Specifies the object to be saved.}

\item{envir}{Environment containing the \code{obj}. (default: \code{.GlobalEnv}).}
}
\value{
Returns the object after loading.
}
\description{
Loading and saving data.
}
\examples{

\dontrun{# Load an object
pcy <- loadData("pcy", tech="TECH")}
\dontrun{# Save an object 
data(pcy)
saveData(pcy, tech="TECH")}
}
