% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfGraphics.R
\name{mapPlotPar}
\alias{mapPlotPar}
\title{mapPlot Parameters}
\usage{
mapPlotPar(...)
}
\arguments{
\item{...}{Numeric, character or function. 

Map types:
\itemize{
  \item{ RB }{ Representation Bin map. }
  \item{ B }{ Bin map. }
  \item{ RBB }{ Representation Binary Bin map. }
  \item{ BB }{ Binary Bin map. }
  \item{ RH }{ Representation Heatmap.}
  \item{ H }{  Heatmap. }
  \item{ RSITE }{ Representation site map.}
  \item{ SITE }{ Site map. }
  \item{ RBSITE }{ Representation Bin site map. }
  \item{ BSITE }{ Bin site map. }
}
Root parameters:
\itemize{
  \item{ \code{breaks} }{ Breaking intervals per map type. }
  \item{ \code{colors} }{ Colors per map type. }
  \item{ \code{legend} }{ Legend display per map type. }
  \item{ \code{frame} }{ Frame display per map type. }
  \item{ \code{mar} }{ Chart margins (default: 5.1, 3.1, 4.1, 3.1). }
}
\code{breaks} parameters:
\itemize{
  \item{ \code{breaks.border} }{ Map border break intervals (default: -5.0, -4.0, -2.1, -2.0, -1.0). }
  \item{ \code{breaks.mapType} }{ Break intervals per map type. }
  \item{ \code{breaks.tr} }{ Heatmap zero level threshold. }
}
\code{colors} parameters: 
\itemize{
  \item{ \code{colors.n} }{ Number of color levels (default: 10). }
  \item{ \code{colors.bin0} }{ Good die color. }
  \item{ \code{colors.border} }{ Border colors. }
  \item{ \code{colors.mapType} }{ Colors per map type. }
}}
}
\description{
Gets, sets and resets visual aspects of mapPlot.
}
\details{
Use parameter names to get values, use assignment to set parameter
values and use keyword \code{default} to load default values.
}
\examples{

data(wbm)

# Plot a wbm corresponding to the fifth row
mapPlot(wbm[5])
# Plot a wbm specified by lot and wafer
mapPlot(wbm[lot=="LOT001"&wafer==12]) 

\dontrun{# Plot multiple maps consecutively
# Use console for browsing
mapPlot(wbm[1:10])
}
# Plot maps next to each other for comparison by matching lot and wafer
mapPlot(wbm[5], wbm) 

# Show probe date below and turn off legend
mapPlot(wbm[lotWafer=="LOT001-W12"], subMain = "probeEndDate", showLegend = FALSE) 

# Advanced: Add a point to wbm
mapPlot(wbm[40], addPoints=matrix(c(10,11,99), 1,3, byrow=TRUE))

# Print all visual parameters
mapPlotPar()

# Get chart margins
mapPlotPar(mar)

# Set chart margins
mapPlotPar(mar=c(5.2, 3.2, 4.2, 3.2))

# Set number of heatmap color levels
mapPlotPar(colors.n=5)

# Reload default visual parameters
mapPlotPar(default)
}
