% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semiFrame.R
\name{mapType.semiFrame}
\alias{mapType.semiFrame}
\alias{mapType}
\alias{mapType<-.semiFrame}
\title{Get or Set Map Type Attribute}
\usage{
\method{mapType}{semiFrame}(obj)

\method{mapType}{semiFrame}(obj) <- value
}
\arguments{
\item{obj}{\code{semiFrame} object.}
}
\value{
\code{mapType} attribute of the \code{semiFrame} object.
}
\description{
Gets or sets \code{mapType} attribute.
}
\details{
See \code{\link{mapPlotPar}} for more information on map types.
}
\examples{

data(wbm)

# Get map type
mapType(wbm)

}
