% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semiFrame.R
\name{modelM.semiTable}
\alias{modelM.semiTable}
\alias{modelM}
\alias{modelM.semiFrame}
\title{Get Model Output}
\usage{
\method{modelM}{semiTable}(obj)

\method{modelM}{semiFrame}(obj, data = "last")
}
\arguments{
\item{obj}{\code{semiTable} or \code{semiFrame} object.}

\item{data}{Character. A \code{semiFrame} object can store models on both parametric and map data. Choose
\itemize{
  \item{ "par" }{ for model on parameters. }
  \item{ "map" }{ for model on maps. }
  \item{ "last" }{ for last modified model. }
}}
}
\value{
The model specific output.
}
\description{
Gets model output of the \code{\link{model}} function ie. the model itself.
}
\examples{
data(pcy)

# Train CART model
pcy2 <- model(pcy, type="class", model="cart")

# Get the model itself
modelM(pcy2)

}
