% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfGraphics.R
\name{modelPlotPar}
\alias{modelPlotPar}
\title{Model Plot Parameters}
\usage{
modelPlotPar(...)
}
\arguments{
\item{...}{Numeric, character or function. 

Root parameters
\itemize{
  \item{ \code{heat} }{ Heatmap color scheme. }
  \item{ \code{colors} }{ Color schemes. }
  \item{ \code{cexAxis} }{ Axis cex for corresponding chart type. }
}
\code{heat} parameters:
\itemize{
  \item{ heat.n }{ Number of color shades (default: 11). }
  \item{ heat.breaks }{ Intervals breaks. }
  \item{ heat.colors }{ Shading colors. }
}
\code{colors} parameters:
\itemize{
  \item{ colors.cart.default.box }{ CART default box color (default: 0). }
  \item{ colors.cart.default.border }{ CART default border color (default: "black"). }
  \item{ colors.cart.br2.box }{ CART box color in \code{color="majority"} mode (default: "#fdae61", "#a6d96a"). }
  \item{ colors.cart.br2.border }{ CART border color in \code{color="majority"} mode (default: "#fdae61", "#a6d96a"). }
  \item{ colors.dr.box }{ Dimensionality reduction box color (default: "#80b1d3"). }
  \item{ colors.dr.border }{ Dimensionality reduction border color (default: "#739FBEFF"). }
  \item{ colors.dr.line }{ Dimensionality reduction line color (default: "#fb8072"). }
}
\code{axis text magnification} parameters:
\itemize{
  \item{ cexAxis.boost }{ Cex for Generalized Boosted Regression model (default: 0.8). }
  \item{ cexAxis.dr }{ Cex for Dimensionality reduction (default: 0.8). }
}}
}
\description{
Gets, sets and reloads default visual parameters of modelPlot.
}
\details{
Use parameter names to get values, use assignment to set parameter
values and use keyword \code{default} to load default values.
}
\examples{
# modelPlot

data(pcy)

# Train Random forest classification
pcy <- model(pcy, type="class", model="forest", splitRatio=1, formula="ProbeYield~.")

# Default variable importance plot
modelPlot(pcy)
# Partial dependence with passed degree and y limit setting
modelPlot(pcy, chart="plotmo", degree1=c(1,3), degree2=FALSE, ylim=c(0,0.5))

# Train CART classification of 'ProbeYield' (median response split by default) on all parameters
pcy <- model(pcy, type="class", model="cart", splitRatio=1, formula="ProbeYield~.", cp=0.03)

# Default decision tree plot
modelPlot(pcy)
\dontrun{
# Hide extra information in nodes and activate snipping tool
modelPlot(pcy, snip=TRUE, extra=0)
}
# Partial dependence with passed degree and y limit setting
modelPlot(pcy, chart="plotmo", yLim=0.9, degree1=1, degree2=FALSE)

# Print all visual parameters
modelPlotPar()
# Get value of a visual parameter
modelPlotPar(colors)

# Default majority mode color scheme
modelPlot(pcy, color="majority")
# Set box color in majority mode
modelPlotPar(colors.cart.br2.box=c("red", "green"))
# Custom majority mode color scheme
modelPlot(pcy, color="majority")

# Reload default visual parameters
modelPlotPar(default)
}
