% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semiFrame.R
\name{normTable}
\alias{normTable}
\title{Get Normalization Table}
\usage{
normTable(obj, verbose = TRUE)
}
\arguments{
\item{obj}{\code{semiTable} or \code{semiFrame} object.}

\item{verbose}{Logical. If \code{FALSE} turns off function messages (default: \code{TRUE}).}
}
\value{
A \code{data.table}
}
\description{
Returns normalization table if present.
}
\examples{

data(pcy)

# Get normalization table
normTable(pcy)

}
