% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfGraphics.R
\name{overlayMapPlot}
\alias{overlayMapPlot}
\title{Maps Overlays Plotting}
\usage{
overlayMapPlot(
  ...,
  mapAlpha = 0.5,
  mapBorder = NULL,
  mapBorderNum = 1,
  main = NULL,
  numMain = TRUE
)
}
\arguments{
\item{...}{\code{semiFrame} objects. If one object is specified, one composite map is constructed
from all rows. When multiple objects are provided, shading is added to display boundary overlaps.}

\item{mapAlpha}{Numeric. Transparency for overlay (default: 0.5).}

\item{mapBorder}{Numeric of length 4. Specifies map cropping (negative values) or extending (positive values). Elements correspond to the up, right, down and left direction (default: \code{NULL}).}

\item{mapBorderNum}{Numeric. Determines to which map object \code{mapBorder} correction will be applied (default: 1).}

\item{main}{Character. Main title.}

\item{numMain}{Logical. If TRUE index of plotted row is included into the main title.}
}
\value{
Plot.
}
\description{
Overlays maps represented by \code{semiFrame} object rows.
}
\examples{

data(wbm)

# Overlay maps corresponding to rows 1-10 with default transparency
overlayMapPlot(wbm[1:10])

# Map from second argument is shaded in red
overlayMapPlot(wbm[1], wbm[2])

# Print all visual parameters
overlayMapPlotPar()

# Get shading colors for multiple map objects regime
overlayMapPlotPar(colors)

# Set shading colors for multiple map objects regime
overlayMapPlotPar(colors=c("yellow", "blue", "black", "green", "red"))

# Reload default visual parameters
overlayMapPlotPar(default)
}
