% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfGraphics.R
\name{overlayMapPlotPar}
\alias{overlayMapPlotPar}
\title{Map Overlay Plot Parameters}
\usage{
overlayMapPlotPar(...)
}
\arguments{
\item{...}{Numeric, character or function. 

Root parameter
\itemize{
 \item{ \code{colors} }{ Shading colors for multiple map objects regime. }
}}
}
\description{
Gets, sets and reloads default visual parameters of overlayMapPlot.
}
\details{
Use parameter names to get values, use assignment to set parameter
values and use keyword \code{default} to load default values.
}
\examples{

data(wbm)

# Overlay maps corresponding to rows 1-10 with default transparency
overlayMapPlot(wbm[1:10])

# Map from second argument is shaded in red
overlayMapPlot(wbm[1], wbm[2])

# Print all visual parameters
overlayMapPlotPar()

# Get shading colors for multiple map objects regime
overlayMapPlotPar(colors)

# Set shading colors for multiple map objects regime
overlayMapPlotPar(colors=c("yellow", "blue", "black", "green", "red"))

# Reload default visual parameters
overlayMapPlotPar(default)
}
