% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfModel.R
\name{peTable}
\alias{peTable}
\title{Process Equipment Table}
\usage{
peTable(obj)
}
\arguments{
\item{obj}{\code{semiTable} or \code{semiFrame} object.}
}
\value{
Process equipment table.
}
\description{
Creates process equipment table from lot equipment history data by counting 
usage for each equipment thus transforming per stage view into per equipment view.
}
