% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfGraphics.R
\name{plotPar}
\alias{plotPar}
\title{General Plot Parameters}
\usage{
plotPar(...)
}
\arguments{
\item{...}{Numeric, character or function. 
Root parameters:
\itemize{
  \item{ \code{black} }{ Black color shade (default: "#556167"). }
  \item{ \code{par} }{ Graphical parameters to be set before plotting. }
  \item{ \code{parSave} }{ Graphical parameters to be saved during plotting. }
}
\code{par} parameters to be set before plotting:
\itemize{
  \item{ \code{par.col} }{ Default plotting color (default: "#556167"). }
  \item{ \code{par.col.axis} }{ Color for axis annotation (default: "#556167"). }
  \item{ \code{par.col.lab} }{ Color for x and y labels (default: "#556167"). }
  \item{ \code{par.col.main} }{ Color for titles (default: "#556167"). }
  \item{ \code{par.col.sub} }{ Color for subtitles (default: "#556167"). }
  \item{ \code{par.col.fg} }{ Plot foreground color (default: "#556167"). }
  \item{ \code{par.col.bty} }{ Color for type of box which is drawn around plots (default: "o"). }
}
\code{parSave} graphical parameters to be saved during plotting, by default these are:
\itemize{
  \item{ "las" }{ Style of axis labels: parallel(=0), perpendicular(=2), horizontal(=1). }
  \item{ "cex.axis" }{ Annotation magnificiation of axis. }
  \item{ "mar" }{ Chart margins. }
  \item{ "mai"}{ Margin size specified in inches. }
  \item{ "xpd" }{ If \code{FALSE}, all plotting is clipped to the plot region, if \code{TRUE}, all plotting is clipped to the figure region, and if \code{NA}, all plotting is clipped to the device region. }
  \item{ "col" }{ Default plotting color. }
  \item{ "col.axis" }{ Color for axis annotation. }
  \item{ "col.lab" }{ Color for x and y labels. }
  \item{ "col.main"}{ Color for titles. }
  \item{ "col.sub" }{ Color for subtitles. }
  \item{ "fg" }{ Plot foreground color. }
  \item{ "bty" }{ Color for type of box which is drawn around plots. }
}}
}
\description{
Gets, sets and resets global visual aspects of plots.
}
\details{
Use parameter names to get values, use assignment to set parameter
values and use keyword \code{default} to load default values.
}
\examples{

# Print all visual parameters
plotPar()

# Get value of a visual parameter
plotPar(black)

# Get value of a graphical parameters to be set before plotting
plotPar(par)

# Set axis color to red
plotPar(par.col.axis="red")

# Reload default visual parameters
plotPar(default)

}
