% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfTools.R
\name{remChar}
\alias{remChar}
\title{Removing character parts}
\usage{
remChar(
  obj,
  ...,
  remLeft = 0,
  remRight = 0,
  keep = 0,
  charLen = NULL,
  inRegex = NULL,
  exRegex = NULL,
  regexCol = NULL,
  useDataType = c("PC", "MET", "LEH", "DEF"),
  col = NULL
)
}
\description{
Removing character parts, from left or from right
}
