% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfTools.R
\name{remColNA}
\alias{remColNA}
\title{Remove Columns with NAs}
\usage{
remColNA(obj, ratio = 0.9, verbose = TRUE)
}
\arguments{
\item{obj}{\code{semiTable} or \code{semiFrame} object.}

\item{ratio}{Specifies threshold for column removal with NA values (default: 0.9).}

\item{verbose}{Logical. If \code{FALSE} turns off function messages (default: \code{TRUE}).}
}
\value{
Copy of the \code{semiTable} or \code{semiFrame} object.
}
\description{
Removes columns with NAs using a cutoff ratio.
}
\examples{

data(pcy)

pcy2 <- variable(pcy, "myPARAM", NA ,"PC")
pcy2 <- remColNA(pcy2)

}
