% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfTools.R
\name{rename}
\alias{rename}
\title{Rename Variables}
\usage{
rename(obj, old, new)
}
\arguments{
\item{obj}{\code{semiTable} or \code{semiFrame} object.}

\item{old}{Character. Specifies a single variable name or group of variable names to be renamed.}

\item{new}{Character. Specifies new variable name or names to be set correspondingly.}
}
\value{
Copy of the \code{semiTable} or \code{semiFrame} object.
}
\description{
Renames variables (columns).
}
\examples{

data(pcy)

# Set the cluster name from "clear" to "clear2"
pcy2 <- rename(pcy, "clear", "clear2")

# Rename multiple variables
pcy2 <- rename(pcy, c("clear", "center", "PARAM01"), c("clear2", "center2", "PARAM99"))
 
}
