% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfGraphics.R
\name{rocPlot}
\alias{rocPlot}
\title{ROC Plot}
\usage{
rocPlot(
  obj,
  chart = "train",
  main = NULL,
  data = "last",
  modelThr = 0.5,
  colorize = TRUE,
  text.adj = c(-0.2, 1.7),
  ...
)
}
\arguments{
\item{obj}{\code{semiTable} or \code{semiFrame} object.}

\item{chart}{Character. Specifies if "train" (default) or "test" data are used for ROC curve.}

\item{main}{Character. Main title.}

\item{data}{Character. A \code{semiFrame} object can store models on both parametric and map data. Choose
\itemize{
  \item{ "par" }{ for model on parameters. }
  \item{ "map" }{ for model on maps. }
  \item{ "last" }{ for last modified model. }
}}

\item{modelThr}{Numeric. Specified threshold will be labeled (default: 0.5).}

\item{colorize}{Logical. Determines whether the ROC curve should be colorized according to
threshold (default: \code{TRUE}).}

\item{text.adj}{Numeric. Specifies adjustment of the labels (default: c(-0.2, 1.7), ...)).}

\item{...}{Character. Additional parameters passed to \code{ROCR::plot} function.}
}
\value{
Plot.
}
\description{
Plots ROC curve for binary classification models stored in \code{semiTable} or \code{semiFrame} objects.
}
\examples{

data(pcy)

# Random forest data
pcy <- model(pcy, type="class", model="forest", splitRatio=0.8, modelFormula="ProbeYield~.")

# Train data used by default
rocPlot(pcy)

# Test data used, model threshold 0.7 labeled 
rocPlot(pcy, chart="test", modelThr=0.7)

}
