% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfGraphics.R
\name{sitePlot}
\alias{sitePlot}
\title{Site Wafer Map Plot}
\usage{
sitePlot(
  obj,
  ...,
  color = "heat",
  fun = median,
  main = NULL,
  numMain = TRUE,
  siteMap = NULL,
  nbrhdIdx = NULL,
  sitePoints = FALSE,
  siteMapSignal = NULL,
  siteCoor = NULL,
  autoNorm = TRUE,
  useDataType = sfUseDataType,
  col = NULL
)
}
\arguments{
\item{obj}{\code{semiTable} object.}

\item{...}{Data columns to be used. Each column can be entered in following forms:
\itemize{
  \item{ integers } { Specify column indexes in the column group given by
    \code{useDataType} parameter. }
  \item{ characters }{ Specify column names. Both character vectors or strings without quotes
    can be used. } 
} 
Browsing mode is activated if multiple columns are provided.}

\item{color}{Character. Color mode:
\itemize{
  \item{ "heat" } { Darkest neighborhood color corresponds to highest parameter center value (default). }
  \item{ "single" }{ Single color is used for all neighborhood rectangles. } 
}}

\item{fun}{Function. Centering function applied on data columns per site. For example
\code{median} (default), \code{mean}, \code{sd}.}

\item{main}{Character. Main title.}

\item{numMain}{Logical. When \code{TRUE} index of plotted column within corresponding \code{useDataType}
is included into the main title.}

\item{siteMap}{Advanced: \code{semiFrame}. Specifies underlying wafer map (default: \code{NULL}).}

\item{nbrhdIdx}{Advanced: Matrix. Specifies site neighborhood indexes (default: \code{NULL}).}

\item{sitePoints}{Advanced: Logical. If \code{TRUE} shows site dice (default: \code{FALSE}).}

\item{siteMapSignal}{Advanced: Character. Specifies column that determines which siteMap
is going to be used if multiple siteMaps are provided (default: \code{NULL}).}

\item{siteCoor}{Advanced: List. User defined site coordinates.}

\item{autoNorm}{Logical. If \code{TRUE}, data prior normalization is restored for plotting.}

\item{useDataType}{Character. Specifies \code{obj} column group for integer column indexing.
First found match is used when multivalued vector is provided. 
See \code{\link{dataType}} for more information on data types.}

\item{col}{Character. Direct column input using character column names or integer column indices (default: \code{NULL}).}
}
\value{
If no data columns are provided, site statistics are returned. Plot otherwise.
}
\description{
Plots parameters distribution centers (medians by default) per site. 
Darkest neighborhood color corresponds to highest parameter center value.
Site yield values are determined by counting good dice in the neighborhood rectangles.
}
\examples{

data(pcy)

# Plot siteid per site with default heatmap color scheme.
sitePlot(pcy, siteid)
# Plot siteid per site with single color scheme.
sitePlot(pcy, siteid, color="single")

\dontrun{# Various ways to input variables for plotting
# Use console for browsing
var <- 10
sitePlot(pcy, 1, c(1,5), c(1:5), c(var), c("siteid"), 1:5, 8:-1, siteid, PARAM01:last)
}
# Plot PCMYield mean
sitePlot(pcy, PCMYield, fun=mean)

# Print all visual parameters
sitePlotPar()

# Get significant digits setting
sitePlotPar(text.signif)
# Set number of significant digits
sitePlotPar(text.signif=5)

# Reload default visual parameters
sitePlotPar(default)
}
