% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfGraphics.R
\name{sitePlotPar}
\alias{sitePlotPar}
\title{Site Plot Parameters}
\usage{
sitePlotPar(...)
}
\arguments{
\item{...}{Numeric, character or function. 

Root parameter:
\itemize{
 \item{ \code{text} }{ Site text parameters. }
}
\code{text} parameters:
\itemize{
  \item{ \code{text.cex} }{ Text magnification (default: 1.5). }
  \item{ \code{text.pos} }{ Text relative position. Defaults to 1 i.e. below coordinates. }
  \item{ \code{text.offset} }{ Text offset from relative position (default: 0.4). }
  \item{ \code{text.flatPos} }{ Text relative position for site near flat if too close to the edge. Defaults to 3 i.e. above coordinates. }
  \item{ \code{text.flatOffset} }{ Text offset from relative position for site near flat if too close to the edge (default: 0.5). }
  \item{ \code{text.signif} }{ Number of significant digits (default: 3). }
  \item{ \code{text.colors} }{ Text shading scheme. }
}
\code{text.colors} parameters:
\itemize{
  \item{ \code{text.colors.single} }{ Text color in single color mode (default: "black"). }
  \item{ \code{text.colors.n} }{ Number of text shades (default: 4). }
  \item{ \code{text.colors.breaks} }{ Shading intervals breaks. }
  \item{ \code{text.colors.br} }{ Shading colors. }
}}
}
\description{
Gets, sets and reloads default visual parameters of sitePlot.
}
\details{
Function \code{mapPlot} is used for wafer rendering. See \code{?mapPlotPar} for heatmap
color scheme settings.

Use parameter names to get values, use assignment to set parameter
values and use keyword \code{default} to load default values.
}
\examples{

data(pcy)

# Plot siteid per site with default heatmap color scheme.
sitePlot(pcy, siteid)
# Plot siteid per site with single color scheme.
sitePlot(pcy, siteid, color="single")

\dontrun{# Various ways to input variables for plotting
# Use console for browsing
var <- 10
sitePlot(pcy, 1, c(1,5), c(1:5), c(var), c("siteid"), 1:5, 8:-1, siteid, PARAM01:last)
}
# Plot PCMYield mean
sitePlot(pcy, PCMYield, fun=mean)

# Print all visual parameters
sitePlotPar()

# Get significant digits setting
sitePlotPar(text.signif)
# Set number of significant digits
sitePlotPar(text.signif=5)

# Reload default visual parameters
sitePlotPar(default)
}
