% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfTools.R
\name{siteYield}
\alias{siteYield}
\title{Calculates and transfers site yield from maps}
\usage{
siteYield(
  from,
  to,
  fromBin = "all",
  allName = "PCMYield",
  binName = NULL,
  binNameRoot = allName,
  digits = 2,
  siteMap = NULL,
  nbrhdIdx = NULL,
  siteMapSignal = NULL
)
}
\description{
Calculates and transfers site yield from maps
}
