% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfModel.R
\name{summary.semiFrame}
\alias{summary.semiFrame}
\alias{summary}
\alias{summary.semiTable}
\title{Model Summary}
\usage{
\method{summary}{semiFrame}(
  obj,
  data = "last",
  details = TRUE,
  special = "none",
  modelThr = 0.5,
  clusterThr = 0.5,
  give = FALSE,
  first = NULL,
  positiveIdx = 2
)

\method{summary}{semiTable}(...)
}
\arguments{
\item{obj}{\code{semiTable} or \code{semiFrame} object.}

\item{data}{Character. A \code{semiFrame} object can store models on both parametric and map data. Choose
\itemize{
  \item{ "par" }{ for model on parameters. }
  \item{ "map" }{ for model on maps. }
  \item{ "last" }{ for last modified model. }
}}

\item{details}{Logical. Specifies summary output extent. If \code{TRUE}, detailed function specific output is returned (default: \code{TRUE}).}

\item{special}{Character. Defines the output to be printed out. If \code{special="cp"}, cross-validation error and standard deviation
together with complexity parameter is depicted. If \code{special="importance"}, CART variable
importance is shown (default: "none").}

\item{modelThr}{Numeric. Specifies the model threshold (default: 0.5).}

\item{clusterThr}{Numeric. Threshold for cluster membership (default: 0.5).}

\item{give}{Logical. If \code{TRUE} returns statistics (default: \code{FALSE}).}

\item{first}{Integer. Specifies first \code{n} values of importance variables for the \code{pca} model (default: \code{NULL}).}

\item{positiveIdx}{Integer. In case of binary classifications defines the positive level to build the confusion table. Possible values 1 or 2 (default: 2).}
}
\value{
Argument specific summary output with result summaries.
}
\description{
Produces various model summaries.
}
\examples{

data(wbm)
data(pcy)

# Show cluster membership summary together with particular bin values
summary(wbm)

# Train CART classification of 'ProbeYield' (median response split by default) on all parameters
pcy2 <- model(pcy, type="class", model="cart", splitRatio=1, formula="ProbeYield~.", cp=0.03)

# Get summary with cross-validation error, standard deviation and complexity parameter information
summary(pcy2, special="cp")

}
