% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semiFrame.R
\name{tab.semiTable}
\alias{tab.semiTable}
\alias{tab<-.semiTable}
\title{Get or set table attribute}
\usage{
\method{tab}{semiTable}(obj, dataType = NULL, addRow = FALSE)

\method{tab}{semiTable}(obj) <- value
}
\description{
semiTable method to get or set table attribute
}
