% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfModel.R
\name{unfoldCluster}
\alias{unfoldCluster}
\title{Multiclass Cluster Column Unfold}
\usage{
unfoldCluster(obj, col, labels = "clust", add = TRUE)
}
\arguments{
\item{obj}{\code{semiTable} or \code{semiFrame} object.}

\item{col}{Character. Defines column name of the categorical variable to unfold into separate cluster variables.}

\item{labels}{Character. Either single name root for new variables or vector of labels for each category (default: "clust").}

\item{add}{Logical. If \code{TRUE} newly spawned variables will be added to the former ones, if \code{FALSE} former cluster variables will be removed (default: \code{TRUE}).}
}
\value{
Copy of the semiTable or semiFrame object with unfolded cluster variables.
}
\description{
Unfolds multiclass cluster column into single class cluster columns.
}
\examples{

data(pcy)

# Create a multi class cluster variable
pcy2 <- variable(pcy, "multiClass", c("A", "B", "C")[round(runif(nrow(pcy), 1, 3))], "C")
# Unfold a multi class variable into regular cluster variables
unfoldCluster(pcy2, "multiClass", add=TRUE)

}
