% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfTools.R
\name{unfoldCol}
\alias{unfoldCol}
\title{Unfold Column}
\usage{
unfoldCol(
  obj,
  col,
  method = "merge",
  key = "default",
  dataType = NULL,
  targetDataType = NULL,
  discardCol = NULL,
  moveCol = NULL,
  sep = "_"
)
}
\arguments{
\item{obj}{\code{semiTable} or \code{semiFrame} object.}

\item{col}{Character. Column name of the categorical variable to unfold.}

\item{method}{Character. Defines method to be used for unfolding: "merge" (default) - only cases with all the categories covered will be kept, 
"add" - all cases will be kept with the missing categories filled as \code{NA}.}

\item{key}{Character. Specifies merging key to be used for unfolding (default: "default").}

\item{dataType}{Character. Data type of variables that will be used for unfolding. 
Each variable will spawn a new variable for each category (default: \code{NULL}).}

\item{targetDataType}{Character. Data type of newly spawned variables. If not specified the same data type will be applied (default: \code{NULL}).}

\item{discardCol}{Character. Specifies column name to be discarded (default: \code{NULL}).}

\item{moveCol}{Character. Specifies column name to be moved before the unfolded columns (default: \code{NULL}).}

\item{sep}{Character. Specifies separator used for new column name creation (default: "_").}
}
\value{
Copy of the semiTable or semiFrame object with new variable groups.
}
\description{
Unfolds a column with categorical variable horizontally into groups of variables.
}
\examples{

data(pcy)

# Unfold siteid column to create a group of PC params for each site
pcyu <- unfoldCol(pcy, "siteid", method="merge", key="lotWafer", dataType="PC", discardCol=c("siteid", "lotWaferSiteid"))

}
