% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfTools.R
\name{variable}
\alias{variable}
\title{Manage Variables}
\usage{
variable(obj, varName, varData, dataType = NULL)
}
\arguments{
\item{obj}{\code{semiTable} or \code{semiFrame} object.}

\item{varName}{Character. Specifies variable's name. The variable will be added if not present.}

\item{varData}{Vector. Specifies variable's data. Use \code{NULL} to remove a variable.}

\item{dataType}{Character. Defines the data type of a variable. Mandatory for new variables. 
See \code{\link{dataType}} for more information on data types. (default: \code{NULL})}
}
\value{
Copy of the \code{semiTable} or \code{semiFrame} object.
}
\description{
Adds, modifies or removes a variable in a \code{semiTable} or \code{semiFrame} object.
}
\examples{

data(pcy)

# Introduce a new variable myPARAM
pcy2 <- variable(pcy, "myPARAM", 2*pcy[,PARAM02]-pcy[,PARAM03]/2, "PC")

# Modify myPARAM variable
pcy2 <- variable(pcy2, "myPARAM", 2*pcy[,PARAM02])

# Remove 'myPARAM'
pcy2 <- variable(pcy2, "myPARAM", NULL)


}
\seealso{
\code{\link{dataType}} for available data types.
}
