% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semiFrame.R
\name{view.data.table}
\alias{view.data.table}
\title{View data.table}
\usage{
\method{view}{data.table}(
  obj,
  n = 5L,
  options = list(pageLength = 25),
  class = "display nowrap",
  ...
)
}
\description{
View data.table
}
