% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semiFrame.R
\name{view.semiFrame}
\alias{view.semiFrame}
\title{View semiFrame}
\usage{
\method{view}{semiFrame}(
  obj,
  n = 5L,
  options = list(pageLength = 25),
  class = "display nowrap",
  ...
)
}
\description{
View semiFrame
}
