% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semiFrame.R
\name{view.semiTable}
\alias{view.semiTable}
\title{View semiTable}
\usage{
\method{view}{semiTable}(
  obj,
  n = 5L,
  options = list(pageLength = 25),
  class = "display nowrap",
  ...
)
}
\description{
View semiTable
}
