% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfGraphics.R
\name{xyPlot}
\alias{xyPlot}
\title{Scatter Plot}
\usage{
xyPlot(
  obj,
  ...,
  outlier = NULL,
  z = c("ProbeYield", "UPYield"),
  zBr = 10,
  zBrMethod = "yieldHeat",
  zLevel = NULL,
  zOutlier = NULL,
  main = NULL,
  linFit = FALSE,
  numMain = TRUE,
  showLegend = TRUE,
  yLog = FALSE,
  xLim = NULL,
  yLim = NULL,
  autoNorm = TRUE,
  drawOrder = NULL,
  useDataType = sfUseDataType,
  col = NULL
)
}
\arguments{
\item{obj}{\code{semiTable} or \code{semiFrame} object.}

\item{...}{Data columns to be used. Each column can be entered in following forms:
\itemize{
  \item{ integers } { Specify column indexes in the column group given by
    \code{useDataType} parameter. }
  \item{ characters }{ Specify column names. Both character vectors or strings without quotes
    can be used. } 
} 
First column specifies x axis. Second column specifies y axis. Browsing mode is activated if multiple 
y axis columns are provided.}

\item{outlier}{Numeric. Outlier filtering of data column before plotting. Outliers are data points outside the
interval <(Q1-k1*IQR),(Q3+k2*IQR)>, where Q1 and Q3 are quartiles, IQR is interquartile range and
k1 and k2 are values specified by \code{outlier} parameter as its first and second element. 
If single value is provided, k1 = k2 setting is used. No outlier filtering is applied by default.}

\item{z}{Vector. Specifies column for \code{z} axis. This column can be broken into intervals 
using \code{zBr} and \code{zBrMethod} for numerical variables and using \code{zLevel} for 
categorical variables. Subsets of data column corresponding to \code{z} column breaks are 
then plotted with different colors. Input methods:
\itemize{
  \item{ integers } { Specify column indexes in the column group given by
    \code{useDataType} parameter. }
  \item{ characters }{ Specify column names. } 
} 
First found match is used when multivalued vector is provided.}

\item{zBr}{Vector. Breaks of a numeric \code{z} column, works in conjunction with \code{zBrMethod} 
setting. Single integer value determines number of breaks. Vector of \code{length>1} can be numerical,
character or POSIXct and represents interval breaking points. Special character values "min", "mean",
"median", "max" and "seq" can be used.}

\item{zBrMethod}{Character. Relevant for numeric \code{z} columns, works in conjunction with \code{zBr} setting. 
Split methods:
\itemize{
  \item{ "median" }{ Breaks \code{z} column with \code{median}. }
  \item{ "mean" }{ Breaks \code{z} column with \code{mean}. }
  \item{ "thr" }{ Breaks \code{z} column with threshold value specified by \code{zBr}. }
  \item{ "yieldHeat" }{ Corresponds to \code{zBr=c("seq(min, median, length.out=brValue)", "max")}, 
   where \code{brValue} is the value specified by \code{zBr}. }
  \item{ "linear" }{ Breaks \code{z} column into equidistant intervals, number of splits is given
   by \code{zBr}. }
  \item{ "hist" }{ Breaks \code{z} column using graphic's \code{hist} function. Values
   are passed to hist's \code{breaks} parameter using \code{zBr}. }
}}

\item{zLevel}{Vector. Non-null value indicates a categorical \code{z} column. Specifies which categories to use
for split. Use special value "all" to include all categories.}

\item{zOutlier}{Numeric. Outlier filtering of \code{z} column before plotting. For more details 
see \code{outlier} parameter.}

\item{main}{Character. Main title.}

\item{linFit}{Logical. Adds linear fit to the chart (default: \code{FALSE}).}

\item{numMain}{Logical. When \code{TRUE} index of plotted column within corresponding \code{useDataType}
is included into the main title.}

\item{showLegend}{Logical. If \code{TRUE} legend is displayed.}

\item{yLog}{Logical. If \code{TRUE} y axis is depicted in logarithmic scale.}

\item{xLim}{Numeric. User defined x axis limits. Vector of length 2.}

\item{yLim}{Numeric. User defined y axis limits. Vector of length 2.}

\item{autoNorm}{Logical. If \code{TRUE}, data prior normalization is restored for plotting.}

\item{drawOrder}{Integer. Specifies drawing order.}

\item{useDataType}{Character. Specifies \code{obj} column group for integer column indexing.
First found match is used when multivalued vector is provided. 
See \code{\link{dataType}} for more information on data types.}

\item{col}{Character. Direct column input using character column names or integer column indices (default: \code{NULL}).}
}
\value{
Plot.
}
\description{
Plots xy scatter plot of a \code{semiTable} or \code{semiFrame} data column. Chart colors are based
on \code{z} axis column.
}
\examples{

data(pcy)

# Plot 1st vs 5th variable from the group
xyPlot(pcy, 1, 5)

\dontrun{# Various ways to input variables for plotting, first variable is used for x axis 
# Use console for browsing
var <- 10
xyPlot(pcy, 1, c(1,5), c(1:5), c(var), c("PARAM03"), 1:5, 8:-1, PARAM01, PARAM08:last)
}
# Select numerical z variable with input by index
xyPlot(pcy, 5, 10, z=8)
# Select categorical z variable with input by name
xyPlot(pcy, 2, 10, z="siteid", zLevel=c(1,5))
# Special keyword "all" for all levels
xyPlot(pcy, 2, 10, z="siteid", zLevel="all")
# Select date type z variable
xyPlot(pcy, 1, 10, z="pcStartDate", zBr=as.POSIXct(c("2015/10/01", "2016/01/01", "2016/10/01")))

# Break numerical variables into intervals
# No break
xyPlot(pcy, 1, 10, zBr=1)
# br keywords seq, min, median, mean and max together with values
xyPlot(pcy, 1, 10, zBr=c("min","5.6", "seq(20, 80, length.out=3)", "max"))

# Add linear fit line
xyPlot(pcy, 2, 10, linFit=TRUE)

# Change draw order of points
xyPlot(pcy, 1, 10, zBr=2, drawOrder=c(2,1))

# Hide legend
xyPlot(pcy, 1, 10, main="Custom main", showLegend = FALSE)

# Print all visual parameters
xyPlotPar()
# Get value of a visual parameter
xyPlotPar(colors.zBr)
# Set breaking colors to red and green and legend position to top left
xyPlotPar(colors.zBr=c("red", "green"), legend.pos="topleft")

# Reload default visual parameters
xyPlotPar(default)
}
