% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfGraphics.R
\name{xyPlotPar}
\alias{xyPlotPar}
\title{Scatter Plot Parameters}
\usage{
xyPlotPar(...)
}
\arguments{
\item{...}{Numeric, character or function. 

Root parameters:
\itemize{  
  \item{ \code{colors} }{ Color schemes. }
  \item{ \code{pch} }{ Plot points symbols. }
  \item{ \code{linFit} }{ Linear fit line parameters. }
  \item{ \code{legend} }{ Legend parameters. } 
}
\code{colors} parameters:
\itemize{
  \item{ \code{colors.zLevel} }{ Color scheme for categorical variables. }
  \item{ \code{colors.zBr} }{ Color scheme for numerical variables. }
  \item{ \code{colors.alpha.zLevel} }{ Color transparency for categorical variables (default: 0.8). }
  \item{ \code{colors.alpha.zBr} }{ Color transparency for numerical variables (default: 1). }
}
\code{linFit} parameters:
\itemize{
  \item{ \code{linFit.color} }{ Color of linear fit line (default: "#525252B3"). }
  \item{ \code{linFit.lwd} }{ Width of linear fit line (default: 2). }
}
\code{legend} parameters:
\itemize{
  \item{ \code{legend.pos} }{ Legend position (default: "topright"). }
  \item{ \code{legend.signif} }{ Number of significant digits (default: 3). }
  \item{ \code{legend.lty} }{ Lines type of the legend (default: 0). }
  \item{ \code{legend.lwd} }{ Lines width of the legend (default: 1). }
  \item{ \code{legend.cex} }{ Annotation magnification of the legend (default: 0.9). }
  \item{ \code{legend.bty} }{ Type of box that is drawn around the legend (default: "n"). }
  \item{ \code{legend.segLen} }{ The length of lines drawn to illustrate ‘lty’ and/or ‘lwd’
          (in units of character widths) (default: "0"). }
  \item{ \code{legend.pch} }{ Points symbols of the legend (default: 15). }
  \item{ \code{legend.ptCex} }{ A numeric determining the size of the exemplary vertex in the legend box (default: 1.8). }
}}
}
\description{
Gets, sets and reloads default visual parameters of xyPlot.
}
\details{
Use parameter names to get values, use assignment to set parameter
values and use keyword \code{default} to load default values.
}
\examples{

data(pcy)

# Plot 1st vs 5th variable from the group
xyPlot(pcy, 1, 5)

\dontrun{# Various ways to input variables for plotting, first variable is used for x axis 
# Use console for browsing
var <- 10
xyPlot(pcy, 1, c(1,5), c(1:5), c(var), c("PARAM03"), 1:5, 8:-1, PARAM01, PARAM08:last)
}
# Select numerical z variable with input by index
xyPlot(pcy, 5, 10, z=8)
# Select categorical z variable with input by name
xyPlot(pcy, 2, 10, z="siteid", zLevel=c(1,5))
# Special keyword "all" for all levels
xyPlot(pcy, 2, 10, z="siteid", zLevel="all")
# Select date type z variable
xyPlot(pcy, 1, 10, z="pcStartDate", zBr=as.POSIXct(c("2015/10/01", "2016/01/01", "2016/10/01")))

# Break numerical variables into intervals
# No break
xyPlot(pcy, 1, 10, zBr=1)
# br keywords seq, min, median, mean and max together with values
xyPlot(pcy, 1, 10, zBr=c("min","5.6", "seq(20, 80, length.out=3)", "max"))

# Add linear fit line
xyPlot(pcy, 2, 10, linFit=TRUE)

# Change draw order of points
xyPlot(pcy, 1, 10, zBr=2, drawOrder=c(2,1))

# Hide legend
xyPlot(pcy, 1, 10, main="Custom main", showLegend = FALSE)

# Print all visual parameters
xyPlotPar()
# Get value of a visual parameter
xyPlotPar(colors.zBr)
# Set breaking colors to red and green and legend position to top left
xyPlotPar(colors.zBr=c("red", "green"), legend.pos="topleft")

# Reload default visual parameters
xyPlotPar(default)
}
