% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfTools.R
\name{yield}
\alias{yield}
\title{Yield Calculation}
\usage{
yield(
  obj,
  fromBin = "all",
  allName = "ProbeYield",
  binName = NULL,
  binNameRoot = "Yield",
  digits = 2,
  pdpwCol = "pdpw"
)
}
\arguments{
\item{obj}{\code{semiFrame} object.}

\item{fromBin}{Character. Defines a specific bin or bins for yield calculation ie. which
bins constitute bad dice. Use keyword "all" (default) for all the bins.}

\item{allName}{Character. Specifies a variable name for the yield column for the case when \code{fromBin} equals "all" (default: "ProbeYield").}

\item{binName}{Character. User defined variable name for the yield column (default: \code{NULL}).}

\item{binNameRoot}{Character. Determines the root part of the variable name for the yield column (default: "Yield").}

\item{digits}{Integer. Specifies number of digits to keep (default: 2).}

\item{pdpwCol}{Character. Specifies the denominator PDPW column name (default: "pdpw").}
}
\value{
Copy of the \code{semiFrame} object with yield.
}
\description{
Creates yield variable from maps ie. number of good dice divided by PDPW.
}
\examples{

data(wbm)

# Create yield for bins 2 and 5 with default variable name
yield(wbm, c(2,5))

# Create yield for bin 2 with user defined name root and 3 digits rounding
yield(wbm, 2, binNameRoot="Y", digits=3)

}
