% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfModel.R
\name{yieldDecomp}
\alias{yieldDecomp}
\title{Yield Decomposition}
\usage{
yieldDecomp(
  obj,
  yields = list(up = "UPYield", probe = "ProbeYield", probeSys = "ProbeYieldSys",
    probeRnd = "ProbeYieldRnd"),
  foiYield = "FOIYield"
)
}
\arguments{
\item{obj}{\code{semiTable} or \code{semiFrame} object.}

\item{yields}{List. Specifies the names of yield columns: 
\itemize{
  \item{ "up" }{ Unit probe yield column (default: "UPYield"). }
  \item{ "probe" }{ Probe yield column (default: "ProbeYield"). }
  \item{ "probeSys" }{ Systematic yield column (default: "ProbeYieldSys"). }
  \item{ "probeRnd" }{ Random yield column (default: "ProbeYieldRnd"). }
}}

\item{foiYield}{Character. Specifies Final Outgoing Inspection (FOI) yield column (default: "FOIYield").}
}
\value{
Yield decomposition object.
}
\description{
Decomposes unit probe yield into systematic yield, random yield and FOI yield.
}
\examples{

data(pcy)

# Decompose unit probe yield into systematic, random and FOI yield 
# components.
yieldDecomp(pcy, list(up="UPYield", probe="ProbeYield", 
            probeSys="ProbeYieldSys", probeRnd="ProbeYieldRnd"), 
            foiYield="FOIYield")

}
